requirements.txt
blinker==1.6.2
boto3==1.28.35
botocore==1.31.35
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.7
